--- a/src/main/cms/cms_menu_gps_rescue.c
+++ b/src/main/cms/cms_menu_gps_rescue.c
@@ -64,7 +64,7 @@ static uint8_t gpsRescueConfig_yawP;
 static uint8_t gpsRescueConfig_velP, gpsRescueConfig_velI, gpsRescueConfig_velD;
 
 static uint8_t gpsRescueConfig_pitchCutoffHz;
-static uint8_t gpsRescueConfig_imuYawCogGain;
+static uint8_t gpsRescueConfig_imuYawGain;
 
 static const void *cms_menuGpsRescuePidOnEnter(displayPort_t *pDisp)
 {
@@ -81,7 +81,7 @@ static const void *cms_menuGpsRescuePidOnEnter(displayPort_t *pDisp)
     gpsRescueConfig_velD = gpsRescueConfig()->velD;
 
     gpsRescueConfig_pitchCutoffHz = gpsRescueConfig()->pitchCutoffHz;
-    gpsRescueConfig_imuYawCogGain = gpsRescueConfig()->imuYawCogGain;
+    gpsRescueConfig_imuYawGain = gpsRescueConfig()->imuYawGain;
     return NULL;
 }
 
@@ -101,7 +101,7 @@ static const void *cms_menuGpsRescuePidOnExit(displayPort_t *pDisp, const OSD_En
     gpsRescueConfigMutable()->velD = gpsRescueConfig_velD;
 
     gpsRescueConfigMutable()->pitchCutoffHz = gpsRescueConfig_pitchCutoffHz;
-    gpsRescueConfigMutable()->imuYawCogGain = gpsRescueConfig_imuYawCogGain;
+    gpsRescueConfigMutable()->imuYawGain = gpsRescueConfig_imuYawGain;
 
     return NULL;
 }
@@ -121,7 +121,7 @@ const OSD_Entry cms_menuGpsRescuePidEntries[] =
     { "VELOCITY D",        OME_UINT8 | REBOOT_REQUIRED, NULL, &(OSD_UINT8_t){ &gpsRescueConfig_velD, 0, 255, 1 } },
 
     { "SMOOTHING",         OME_UINT8 | REBOOT_REQUIRED, NULL, &(OSD_UINT8_t){ &gpsRescueConfig_pitchCutoffHz, 10, 255, 1 } },
-    { "IMU_GAIN",          OME_UINT8 | REBOOT_REQUIRED, NULL, &(OSD_UINT8_t){ &gpsRescueConfig_imuYawCogGain, 1, 20, 1 } },
+    { "IMU_YAW_GAIN",      OME_UINT8 | REBOOT_REQUIRED, NULL, &(OSD_UINT8_t){ &gpsRescueConfig_imuYawGain, 5, 20, 1 } },
 
     {"BACK", OME_Back, NULL, NULL},
     {NULL, OME_END, NULL, NULL}