--- a/src/main/flight/gps_rescue.c
+++ b/src/main/flight/gps_rescue.c
@@ -633,7 +633,7 @@ static void sensorUpdate(void)
             // There should be a simple relationship between pitch angle (forward velocity) and IMU yaw-Cog time constant
             // If that's true, we should see twice as much time to fix an IMU error, and with twice the radius, at half the IMU gain?
             // this PR lets us accept the user's IMU gain value for this phase, for testing purposes
-            rescueState.sensor.imuYawCogGain = pitchForwardAngle + fminf(groundspeedErrorRatio, 2.5f);
+            rescueState.sensor.imuYawCogGain = pitchForwardAngle + fminf(groundspeedErrorRatio, 4.0f); // max around 5x
         }
     }