This is a PR purely for testing future modifications to GPS Rescue code.  It changes frequently, according to the tests we need to do.  At present it permits IMU error development to test how the rescue code handles IMU errors.

**WARNING: This code reflects the code in PR #12900 (proposal for master) BUT allows IMU disorientation to develop eg while flying backwards !!!**

**CURRENT INTENT: Testing GPS Rescue behaviour *in the presence of a full 180 degree IMU disorientation state*.**

**Please use the `ATTITUDE` debug!**  Values are aligned with 12900.

**Testing should focus on:**

- finding the upper limit of wind speed for a given max rescue angle when there is a 180 degree IMU error.  Up to 50kph wind should be OK at the default angle of 45 degrees.  This code allows IMU errors to develop, so it tests the absolute worst case scenario in wind.  If you have the opportunity, try values above or below 45 degrees on very windy days after generating 180 degree IMU errors.

- determining the effect of the `gps_rescue_imu_yaw_gain` value on windy days with 180 degree IMU errors.  This sets the tightness of the turn at the start when there is a full 180 degree error.  We need to know how this affects the rescue; small values may cause spiralling, larger values may be more successful in higher wind.

**Notes:**

- The sanity check for fly home phase in this PR requires only that the quad is flying towards home, more often than not, with a time-out in this PR of 30s.  After that a switch initiated rescue will 'level out' and float around or descend for a further 20s.  By all means test this out, but ideally, If the rescue is failing, don't wait more than 25s before you undo the 'rescue' switch and take control back.  

- While testing in high 'away' winds, don't let the quad drift so far away that you lose video or Rx link, unless you are sure that the rescue will work.  Always do these tests close to home.

**<UPDATE 13/6/2023>** *Permitted IMU disorientation again*.  Otherwise made code consistent with PR 12900, including debugs.  Modified the zero IMU range to 1.0m/s as per 12900, ie nothing until 1m/s then linear starting at 20% of normal IMU gain up to 2x at 10m/s. Renamed the CLI value that adjusts turn tightness with IMU errors to `gps_rescue_imu_yaw_gain`.  

**<UPDATE 13/6/2023>** Included code that **changes the amount of groundspeed needed to set the IMU**.  Instead of being 'zero' IMU update below 2m/s, then 'full' IMU update above 2m/s, it is now zero below 0.5m/s.  Above 0.5m/s, gain starts at 10% of normal, increasing linearly to 'normal' at 5m/s, and to double normal rate at 10m/s and above.  Pilots should take off and fly dead straight at about 5-10ms until the arrow stabilises, ideally directly into or away from the wind, to set the arrow towards home.  This should take only about 4-5s at decent groundspeed.  The arrow should then reliably point to home for the rest of the flight.  It may lose orientation (point the arrow in the wrong direction) if flown aggressively, or if flow inverted for extended periods, after impacts, or when flown in situations where wind drift, rather than motor thrust, is a significant part of the ground path.  Please report here if you are able to successfully cause an IMU error.

**<UPDATE 12/06/2023>** **Updated to @ledvinap's latest code from PR#12792**, which considers the thrust vector direction when evaluating the course over ground correction.  This prevents IMU errors from developing when the quad is flown backwards, etc.  The only situations that can now cause a 180 degree IMU error require a groundspeed > 2m/s in the opposite direction to the expected thrust vector for more than 5-6s.  This can only happen on windy days, for example, by pitching the nose forward directly into a decent wind, but allowing the quad to drift backwards in exactly the opposite direction to the expected flight path (ie, backwards).  In that case a classic 180 degree IMU error would develop, and that would be handled by the GPS Rescue code as per the video below.

I tested this today on a day with very little wind.  Every rescue was crisp and immediate.  No matter whether I flew backwards, sideways, or even upside down for brief periods, the quad turned immediately in the correct direction.

**This is absolutely fantastic!**  It will provide a very high level of safety for the initial situation of a rescue.  It means there is very little chance of a pre-existing IMU before starting the rescue, and there isn't enough time for a significant IMU error to develop in the fly home phase, even in solid wind.  And if it did, the rescue code should boost the IMU so that whatever error we had would resolve quickly.

Question:  Should I maintain a build wherein we still can generate 180 degree IMU errors, for rescue code testing?

**<UPDATE 11/06/2023>** Fixed a bug that I introduced when refactoring the pitch iTerm code.  We now get smooth-onset of pitch (velocity) iTerm and Roll, and block tiTerm until the groundspeed error is very low.  Net effect is a much smoother initial velocity acquisition period, as can be seen in the video above.  As a result, we now apply the full configured maximum pitch angle limit throughout the rescue, which should make rescues against wind more reliable.  

Also pushed a commit for activate earth referencing of yaw during a rescue, even if the user set it off, to ensure consistent roll response while pitched forward.

The tightness of the initial turn radius can be modified with the `gps_rescue_imu_gain` setting.  This is a work-in-progress parameter.  The default value of 10 gives results as per the above video.  A value of 6-7 will result in tighter turns, and may be useful when flying in smaller areas or for emergency rescue when learning LOS, but may lead to spiralling behaviour due to excessive turning, especially in higher winds.  A value of 13-15 will result in a wider turn at the start, but this *may* be better in higher wind, when combined with a higher than default max angle.

Video of this behaviour here: https://www.youtube.com/watch?v=FKKUgGd9Ocw

I'd be very grateful for people to test the effect of gps rescue max angle, primarily, and of the `gps_rescue_imu_gain` setting, in windy conditions, in order to help determine the most suitable default values (currently 45 degrees and 10, respectively)

<UPDATE 06/06/2023> !IMPORTANT!  Added @ledvinap's new, revised code from PR#12792.  Please test cases of 180 IMU disorientation with the debug mode set to ATTITUDE.

<UPDATE 04/06/2023> Reversed the first version of the code from PR#12792, which caused glitches in IMU output.  Added back a modified dynamic IMU assignment.  Utilised debugs more effectively. 

<UPDATE> 02/06/23 Added @ledvinap's code from his PR#12792.  Testing is needed to confirm that it fixes the issue in the IMU code where at error angles greater than 90 degrees, where the `ez_ef` IMU error parameter would decline towards zero when the error was 180 degrees. Modified the debug slightly to help with this. 

**Background**:  If the IMU gain is fixed at 1.0, the normal value, at the start of such a rescue, it takes a very long time for the IMU to correct itself in the presence of a heading error.  ledvinap identified an issue where an IMU error of 180 degrees resulted in a zero or near-zero IMU gain value than 1.0.  And although 4.4.2 prevents IMU error acquisition during the climb and rotate phases, an IMU error can be pre-existing due to flying backwards (which can cause the dreaded full 180 degree IMU error) or sideways. These factors could result in flyaways that might not correct before a sanity check failure.

We also know that if the IMU gain is too high, relative to the allowed maximum pitch angle of the quad and the drift speed, then the quad can enter a yaw circling behaviour.  

Recent IMU adaptation PR's adjust the IMU gain at different points in the early part of a 180-degree IMU error rescue.  These have improved behaviour.  This PR is a test bed for improvements in the code that handles the early part of a GPS rescue to prevent flyaways and provide more reliable returns, even in the face of full 180 degree errors.

I have temporarily hacked the 'attitude' debug to provide the following:
```
0 = accADC[X] // X axis attitude
1 = accADC[Y] // Y axis attitude
2 = IMU CogYaw gain value * 100
3 = ez_ef value * 100 (after multiplication by IMU CogYaw gain when in a rescue)
4 = velocity to home in cm/s
5 = groundspeedErrorRatio * 100 (200 means the error between groundspeed and speed to home is 20ms)
6 = pitchForwardAngle * 100 (value of 100 means 30 degree pitch forward, 200 means 60 degrees)
7 = dcmKpGain * 100 // for testing dcmKpGain refactor or dcmKpGain code changes
```

The original version of this PR used fixed IMU gain via CLI, for testing purposes.  We established from this that approximately 3 of gain works well, and though 6 is faster, if sustained it will cause spiralling. 

The current version uses a modification of the dynamic IMU code currently in master.  

In its present form, and only while in a rescue:

- if the quad is flying towards home "normally", ie with minimal groundspeed error, the IMU gain is determined only by the pitch angle of the quad, using the `pitchForwardAngle` factor.  This acts as a simple multiplier of the IMU COG gain.  It is limited to the range 0-1, and will be 1.0 at a pitch angle of 30 degrees or greater, up to a max of 2.0 at a pitch angle of 60 degrees, and 0 if pitch is zero or negative.  This essentially permits IMU adaptation only while pitched forward, and adapts the IMU faster at higher pitch angles.

- if there is a discrepancy between GPS groundspeed and velocity to home, the `groundspeedErrorRatio` value increases above zero.  It is calculated in units of 10m/s of error by default.  The user can change this value with the `gps_rescue_imu_gain` factor in the CLI or OSD.  The value will be zero, or close to zero, if the GPS groundspeed matches the velocity to home.  However if, for example, the quad is flying diametrically away from home at 10m/s (i.e flying away at the `gps_rescue_imu_gain` value), the `groundspeedErrorRatio` will be 2.0.  When flying away twice as fast again, the value will be 4.0.

- the `groundspeedErrorRatio` value, when close to zero, will prevent velocity iTerm from accumulating via the a the `velocityItermAttenuator` parameter unless the quad is flying directly towards home, to prevent unwanted velocity iTerm windup at the start of a rescue with significant IMU errors.

- the final COG gain is the simple sum of the `pitchForwardAngle` and `groundspeedErrorRatio` factors, limited to a max gain of about 5.5 (where 1.0 is normal).

Thanks for testing!