@ctzsnooze 2023-06-01
@ledvinap  
I've done some testing now with fixed, higher, IMU gain values.

It seems that a fixed increase in gain, via the `cogYawGain` method, from the GPS code, appears to result in a correction of the yaw error which starts slowly, and then gets stronger and stronger as the error corrects.  

Unfortunately, the rate at which the yaw error corrects seems to get faster and faster, even as the heading of the quad gets closer and closer to the correct value.  This tends to result in a significant overshoot, and sometimes the overshoot can cause it to circle around completely.

Is there any kind of `integration` of the `ez` or IMU yaw factor within the Mahony code which could explain this?
And if it takes time for the incoming `ez` value to get the yaw value to change, is there some alternate way where we can get a more direct increase in the effective IMU adaptation time?

The most successful 'heuristic' solutions have involved boosting the IMU only when the error is extreme, and cutting back very hard on it when the IMU error is starting to fix itself.  But due to this apparent lag in effect, and the difficulty of quantifying the IMU error magnitude itself, the results have been a bit erratic and unreliable, depending a lot on the initial conditions perhaps, or perhaps because there are integrations happening that I haven't considered.

Anyhow I'd appreciate your thoughts on this, if you have time to share them.