@ctzsnooze 2023-06-01
@KarateBrot @ledvinap 

Are these lines : https://github.com/betaflight/betaflight/blob/66734f4574ce90463d4116f3e082fdbdc3474f57/src/main/flight/imu.c#L291-L299

causing a kind of lowpass delay in acquiring the error value into the IMU?

As mentioned above, I am searching for a way to speed up the IMU correction other than by multiplying up the input error, but I honestly don't understand the code very well, especially these lines.

It looks like a kind of averaging or smoothing function whereby the `gz` value becomes, over time, closer to the `ez * dcmKpGain` value.  As far as I can tell, dT is the time in seconds between IMU updates, ie around about 0.01s.  If that's correct, the output value of `gz` is mostly determined by the incoming value of `ez * dcmKpGain`, and an incoming change in `ez` would very rapidly modify `gz`.  

In any case I'd appreciate an analysis of what these lines are supposed to do, and whether they serve a smoothing function, or what the purpose of these manipulations are.  The comments don't seem to match up.