@ledvinap 2023-06-01
> Unfortunately, the rate at which the yaw error corrects seems to get faster and faster, even as the heading of the quad gets closer and closer to the correct value. If the `cogYawGain` factor is low, it takes a very long time to initially react, leading to long-duration flyaways. If set high, it still takes time to respond, initially, but then gets more and more aggressive, and can overshoot the heading target, and sometimes the overshoot can be such that it circles around completely. It very much seems like the correction gets faster and faster, and the yaw rate faster and faster, over time.

Error is calculated as cross product between CoG and heading. Magnitude of cross product is |a|*|b|*sin(theta). With 180deg error, sin ~ 0 and correction is small. See https://github.com/betaflight/betaflight/pull/12792

> Is there any kind of `integration` of the `ez` or IMU yaw factor within the Mahony code which could explain this? 

Mahony uses PID to steer error to zero, so there is integral term.

> And, if it takes time for the incoming `ez` value to accumulate enough that the yaw value starts to change, is there some alternate way where we can get a more direct increase in the effective IMU adaptation time, ie to change earlier, rather than later?

See above
 
> The most successful 'heuristic' solutions have involved boosting the IMU only when the error is extreme, and cutting back very hard on it when the IMU error is starting to fix itself. But due to this apparent lag in effect, and the difficulty of quantifying the IMU error magnitude itself, the results have been a bit erratic and unreliable, depending a lot on the initial conditions perhaps, or perhaps because there are integrations happening that I haven't considered.

Separating heading and yaw may help here, but it will need some code. Yaw time constant can be kept small (not affecting PID loops etc.) and heading gain adapter according to input data quality (tilt angle, gps confidence etc). Yaw will then slowly track estimated heading to prevent drift.