@ctzsnooze 2023-06-01
@ledvinap thanks for that advice...  
Here is a poor quality video showing static IMU gain factor during the return phase of four rescues.
The max angle is 55 degrees and the IMU values are:
- 10, which causes spiralling
- 4, a bit too aggressive when the yaw correction overshoots
- 3, slow at the start and soft towards the end of the correction
- 1, our old default, where while flying away the correction takes so long that I was worried it would sanity check, I then manually turned too home and lost video, so I had to re-initate the rescue, and was lucky then that there was enough IMU orientation to get the quad back.

https://www.youtube.com/watch?v=-8MdNOgCjbw