@ledvinap 2023-06-01
You can try https://github.com/betaflight/betaflight/pull/12792/files

It evaluated both cross (sin) and dot (cos) product. 
|dot| > 0 means that error is < 90deg and old behavior is used. 
|dot| < 0 means error > 90deg and sin() is replaced with sign(cross), keeping gain at 1

Another consequence of existing (and PR) code is that `heading_ef` is projection of quad X axis and it's magnitude is factored into error. That means that IMU gain is decreasing with tilt angle (great for plane, bad for quad). 
It is possible to normalize `heading_ef` (iNav does it). 
Or better, use projection of Up axis - zero Yaw updating when level (both normal and upside down), maximum at 90deg, both roll and pitch are taken into account