@ctzsnooze 2023-06-01
@ledvinap   Thanks for the reference to PR#12791!  I will put that code into this PR in a separate branch, and test it that way.  I'll make a video as soon as I can test it next.  I'll log ez_ef properly also.  If it works, many, many GPS Rescue pilots will be super grateful!  

Your other comment about tilt angle also explains why when the quad pitches forward harder, it runs wider.  This is also very bad, I'm sure quads were lost from this when the rescue angle was configured to a high value.  

I had seen the quad 'run wide' at higher angles, but thought it was due to the additional velocity that comes from the higher angle was making the arc wider by physics.  However it is definitely undesirable.  If this could be improved, or fixed, then we could use the full rescue angle from the start, which would make rescues in high wind far more reliable.

I know for sure that I couldn't code the suggestions to fix it.  I'd be very grateful if possibly you could draft a PR to test.

Thanks very much indeed for all your help so far!

C